--- app/src/marketplace/components/Modal/SetAllowanceDialog/setAllowanceDialog.pcss
@@ -1,3 +1,0 @@
-.helpText {
-  margin-bottom: 2rem;
-}
