+++ app/src/marketplace/containers/ProductPage/Hero.jsx
@@ -3,12 +3,15 @@
 import React, { useCallback } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { replace } from 'connected-react-router'
+import moment from 'moment'
+import { I18n } from 'react-redux-i18n'
 
 import useProduct from '$mp/containers/ProductController/useProduct'
 import useModal from '$shared/hooks/useModal'
 import usePending from '$shared/hooks/usePending'
 import { selectUserData } from '$shared/modules/user/selectors'
-import { addFreeProduct } from '$mp/modules/purchase/actions'
+import { addFreeProduct } from '$mp/modules/product/services'
+import { getMyPurchases } from '$mp/modules/myPurchaseList/actions'
 
 import ProductDetails from '$mp/components/ProductPage/ProductDetails'
 import HeroComponent from '$mp/components/Hero'
@@ -18,6 +21,8 @@ import {
     selectContractSubscription,
 } from '$mp/modules/product/selectors'
 import { ImageTile } from '$shared/components/Tile'
+import { NotificationIcon } from '$shared/utils/constants'
+import Notification from '$shared/utils/Notification'
 
 import routes from '$routes'
 
@@ -42,12 +47,23 @@ const Hero = () => {
             if (isLoggedIn) {
                 if (isPaid) {
                     // Paid product has to be bought with Metamask
-                    await purchaseDialog.open({
+                    const { started, succeeded, viewInCore } = await purchaseDialog.open({
                         productId,
                     })
                 } else {
                     // Free product can be bought directly
-                    await dispatch(addFreeProduct(productId || ''))
+
+                    // subscribe for one year (TODO: move to constant)
+                    const endsAt = moment().add(1, 'year').unix() // Unix timestamp (seconds)
+
+                    await addFreeProduct(productId || '', endsAt)
+
+                    Notification.push({
+                        title: I18n.t('notifications.productSaved'),
+                        icon: NotificationIcon.CHECKMARK,
+                    })
+
+                    dispatch(getMyPurchases())
                 }
             } else {
                 dispatch(replace(routes.auth.login({
