--- app/src/marketplace/modules/allowance/constants.js
@@ -1,33 +1,0 @@
-// @flow
-
-export const RESET_DATA_ALLOWANCE_STATE: string = 'marketplace/allowance/RESET_DATA_ALLOWANCE_STATE'
-
-export const GET_DATA_ALLOWANCE_REQUEST: string = 'marketplace/allowance/GET_DATA_ALLOWANCE_REQUEST'
-export const GET_DATA_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/GET_DATA_ALLOWANCE_SUCCESS'
-export const GET_DATA_ALLOWANCE_FAILURE: string = 'marketplace/allowance/GET_DATA_ALLOWANCE_FAILURE'
-
-export const SET_DATA_ALLOWANCE_REQUEST: string = 'marketplace/allowance/SET_DATA_ALLOWANCE_REQUEST'
-export const SET_DATA_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/SET_DATA_ALLOWANCE_SUCCESS'
-export const SET_DATA_ALLOWANCE_FAILURE: string = 'marketplace/allowance/SET_DATA_ALLOWANCE_FAILURE'
-export const RECEIVE_SET_DATA_ALLOWANCE_HASH: string = 'marketplace/allowance/RECEIVE_SET_DATA_ALLOWANCE_HASH'
-
-export const RESET_DATA_ALLOWANCE_REQUEST: string = 'marketplace/allowance/RESET_DATA_ALLOWANCE_REQUEST'
-export const RESET_DATA_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/RESET_DATA_ALLOWANCE_SUCCESS'
-export const RESET_DATA_ALLOWANCE_FAILURE: string = 'marketplace/allowance/RESET_DATA_ALLOWANCE_FAILURE'
-export const RECEIVE_RESET_DATA_ALLOWANCE_HASH: string = 'marketplace/allowance/RECEIVE_RESET_DATA_ALLOWANCE_HASH'
-
-export const RESET_DAI_ALLOWANCE_STATE: string = 'marketplace/allowance/RESET_DAI_ALLOWANCE_STATE'
-
-export const GET_DAI_ALLOWANCE_REQUEST: string = 'marketplace/allowance/GET_DAI_ALLOWANCE_REQUEST'
-export const GET_DAI_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/GET_DAI_ALLOWANCE_SUCCESS'
-export const GET_DAI_ALLOWANCE_FAILURE: string = 'marketplace/allowance/GET_DAI_ALLOWANCE_FAILURE'
-
-export const SET_DAI_ALLOWANCE_REQUEST: string = 'marketplace/allowance/SET_DAI_ALLOWANCE_REQUEST'
-export const SET_DAI_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/SET_DAI_ALLOWANCE_SUCCESS'
-export const SET_DAI_ALLOWANCE_FAILURE: string = 'marketplace/allowance/SET_DAI_ALLOWANCE_FAILURE'
-export const RECEIVE_SET_DAI_ALLOWANCE_HASH: string = 'marketplace/allowance/RECEIVE_SET_DAI_ALLOWANCE_HASH'
-
-export const RESET_DAI_ALLOWANCE_REQUEST: string = 'marketplace/allowance/RESET_DAI_ALLOWANCE_REQUEST'
-export const RESET_DAI_ALLOWANCE_SUCCESS: string = 'marketplace/allowance/RESET_DAI_ALLOWANCE_SUCCESS'
-export const RESET_DAI_ALLOWANCE_FAILURE: string = 'marketplace/allowance/RESET_DAI_ALLOWANCE_FAILURE'
-export const RECEIVE_RESET_DAI_ALLOWANCE_HASH: string = 'marketplace/allowance/RECEIVE_RESET_DAI_ALLOWANCE_HASH'
