+++ app/src/marketplace/modules/product/selectors.js
@@ -3,20 +3,17 @@
 import { createSelector } from 'reselect'
 import { denormalize } from 'normalizr'
 import moment from 'moment'
-import BN from 'bignumber.js'
 
 import type { ProductState, StoreState } from '../../flowtype/store-state'
 import type { EntitiesState } from '$shared/flowtype/store-state'
 import type { ProductId, Product, Subscription } from '../../flowtype/product-types'
 import type { StreamIdList, StreamList } from '$shared/flowtype/stream-types'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
-import type { Category } from '../../flowtype/category-types'
+import type { Category } from '$mp/flowtype/category-types'
 import { selectEntities } from '$shared/modules/entities/selectors'
 import { selectSubscriptions } from '../myPurchaseList/selectors'
 import { productSchema, streamsSchema, categorySchema } from '$shared/modules/entities/schema'
-import { isActive } from '../../utils/time'
-import { selectAccountId } from '../web3/selectors'
-import { areAddressesEqual } from '../../utils/smartContract'
+import { isActive } from '$mp/utils/time'
 
 const selectProductState = (state: StoreState): ProductState => state.product
 
@@ -85,13 +82,6 @@ export const selectContractSubscriptionError: (StoreState) => ?ErrorInUi = creat
     (subState: ProductState): ?ErrorInUi => subState.contractSubscriptionError,
 )
 
-export const selectProductIsFree: (state: StoreState) => boolean = createSelector(
-    selectProduct,
-    (product) => (
-        !!product && product.pricePerSecond === '0'
-    ),
-)
-
 export const selectProductIsPurchased: (state: StoreState) => boolean = createSelector(
     selectProductId,
     selectSubscriptions,
@@ -112,35 +102,3 @@ export const selectSubscriptionIsValid: (StoreState) => boolean = createSelector
         productIsPurchased || contractSubscriptionIsValid
     ),
 )
-
-export const selectFetchingProductSharePermission: (StoreState) => boolean = createSelector(
-    selectProductState,
-    (subState: ProductState): boolean => subState.productPermissions.fetchingPermissions,
-)
-
-export const selectProductSharePermission: (StoreState) => boolean = createSelector(
-    selectProductState,
-    (subState: ProductState): boolean => subState.productPermissions.share,
-)
-
-export const selectProductEditPermission: (StoreState) => boolean = createSelector(
-    selectProductState,
-    (subState: ProductState): boolean => subState.productPermissions.write,
-)
-
-export const selectProductPublishPermission = createSelector([
-    selectProduct,
-    selectAccountId,
-    selectProductSharePermission,
-], (product, ownerAddress, canShare): boolean => {
-    const isProductFree = (!!product && (product.isFree !== false || BN(product.pricePerSecond).isEqualTo(0)))
-    if (isProductFree) {
-        return canShare
-    }
-    return canShare && !!(
-        product &&
-        product.ownerAddress &&
-        ownerAddress &&
-        areAddressesEqual(product.ownerAddress, ownerAddress)
-    )
-})
