--- app/src/marketplace/modules/purchase/constants.js
@@ -1,12 +1,0 @@
-// @flow
-
-export const CLEAR_PURCHASE_STATE: string = 'marketplace/purchase/CLEAR_PURCHASE_STATE'
-
-export const BUY_PRODUCT_REQUEST: string = 'marketplace/purchase/BUY_PRODUCT_REQUEST'
-export const BUY_PRODUCT_SUCCESS: string = 'marketplace/purchase/BUY_PRODUCT_SUCCESS'
-export const BUY_PRODUCT_FAILURE: string = 'marketplace/purchase/BUY_PRODUCT_FAILURE'
-export const RECEIVE_PURCHASE_HASH: string = 'marketplace/purchase/RECEIVE_PURCHASE_HASH'
-
-export const ADD_FREE_PRODUCT_REQUEST: string = 'marketplace/purchase/ADD_FREE_PRODUCT_REQUEST'
-export const ADD_FREE_PRODUCT_SUCCESS: string = 'marketplace/purchase/ADD_FREE_PRODUCT_SUCCESS'
-export const ADD_FREE_PRODUCT_FAILURE: string = 'marketplace/purchase/ADD_FREE_PRODUCT_FAILURE'
