--- app/src/marketplace/modules/web3/constants.js
@@ -1,7 +1,0 @@
-// @flow
-
-export const RECEIVE_ACCOUNT: string = 'marketplace/web3/RECEIVE_ACCOUNT'
-export const CHANGE_ACCOUNT: string = 'marketplace/web3/CHANGE_ACCOUNT'
-export const ACCOUNT_ERROR: string = 'marketplace/web3/ACCOUNT_ERROR'
-
-export const UPDATE_ETHEREUM_NETWORK_ID: string = 'marketplace/web3/UPDATE_ETHEREUM_NETWORK_ID'
