--- app/test/unit/modules/allowance/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/allowance/constants'
-
-describe('allowance - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/allowance\//))
-        })
-    })
-})
