--- app/test/unit/modules/purchase/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/purchase/constants'
-
-describe('purchase - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/purchase\//))
-        })
-    })
-})
