--- app/test/unit/modules/web3/constants.test.js
@@ -1,9 +1,0 @@
-import * as constants from '$mp/modules/web3/constants'
-
-describe('web3 - constants', () => {
-    it('is namespaced correctly', () => {
-        Object.keys(constants).forEach((key) => {
-            expect(constants[key]).toEqual(expect.stringMatching(/^marketplace\/web3\//))
-        })
-    })
-})
